backend/requirements.txt
flask
python-dotenv
flask-cors
google-genai
gunicorn==21.2.0